backend/requirements.txt
django>=4.2
djangorestframework
pillow
djoser
djangorestframework-simplejwt
gunicorn
psycopg2-binary
django-cors-headers 